backend/requirements.txt
fastapi
pydantic
pydantic-settings
google-genai
uvicorn[standard]
python-dotenv
sqlalchemy
pdfplumber
psycopg2-binary>=2.9.0
python-multipart
pydantic[email]